pyproject.toml
[tool.poetry]
name = "carbonfootprintapp"
version = "0.1.0"
description = ""
authors = ["Taishi Endo <82435332+EndoTaishi@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.3"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
openai = "^1.14.3"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
